+++ playroom/snippets.tsx
@@ -712,7 +712,7 @@ const cardSnippets: Array<Snippet> = [
     },
     {
         group: 'Cards',
-        name: 'DisplayMediaCard',
+        name: 'DisplayMediaCard with image',
         code: `
         <DisplayMediaCard
           headline={<Tag type="promo">Headline</Tag>}
@@ -737,14 +737,32 @@ const cardSnippets: Array<Snippet> = [
     },
     {
         group: 'Cards',
-        name: 'PosterCard',
+        name: 'DisplayMediaCard with video',
+        code: `
+        <DisplayMediaCard
+          headline={<Tag type="promo">Headline</Tag>}
+          pretitle="Pretitle"
+          title="Title"
+          description="Description"
+          backgroundVideo="https://fr-cert1-es.mytelco.io/2O4-xBJqiMlAfLkseq8RkXs_mv2ACV7Hnt20HqXxNl-mK7KLI3M2dAw"
+          poster="https://source.unsplash.com/900x900/?landscape"
+          button={
+            <ButtonPrimary small href="https://google.com">
+              Action
+            </ButtonPrimary>
+          }
+        />`,
+    },
+    {
+        group: 'Cards',
+        name: 'PosterCard with image',
         code: `
         <PosterCard
           headline={<Tag type="promo">Headline</Tag>}
           pretitle="Pretitle"
           title="Title"
           description="Description"
-          backgroundImage="https://api.lorem.space/image/watch?w=700&h=700"
+          backgroundImage="https://source.unsplash.com/900x900/?landscape"
           onClose={() => {}}
           actions={[
             {
@@ -753,6 +771,24 @@ const cardSnippets: Array<Snippet> = [
               label: "Lightning",
             },
           ]}
+        />`,
+    },
+    {
+        group: 'Cards',
+        name: 'PosterCard with video',
+        code: `
+        <PosterCard
+          headline={<Tag type="promo">Headline</Tag>}
+          pretitle="Pretitle"
+          title="Title"
+          description="Description"
+          backgroundVideo="https://fr-cert1-es.mytelco.io/2O4-xBJqiMlAfLkseq8RkXs_mv2ACV7Hnt20HqXxNl-mK7KLI3M2dAw"
+          poster="https://source.unsplash.com/900x900/?landscape"
+          button={
+            <ButtonPrimary small href="https://google.com">
+              Action
+            </ButtonPrimary>
+          }
         />`,
     },
 ];