+++ src/video.css.ts
@@ -1,9 +1,9 @@
 import {sprinkles} from './sprinkles.css';
 
 export const video = sprinkles({
+    top: 0,
+    left: 0,
     display: 'block',
-    width: '100%',
-    height: '100%',
     background: 'transparent',
     maxWidth: '100%',
     maxHeight: '100%',