+++ src/video.tsx
@@ -1,13 +1,50 @@
 import * as React from 'react';
-import {useMediaBorderRadius} from './image';
+import {ImageContent, ImageError, useMediaBorderRadius} from './image';
 import {AspectRatioElement} from './utils/aspect-ratio-support';
 import {combineRefs} from './utils/common';
 import {getPrefixedDataAttributes} from './utils/dom';
-import {isRunningAcceptanceTest, isSafari} from './utils/platform';
+import {isRunningAcceptanceTest} from './utils/platform';
 import * as styles from './video.css';
 import {vars} from './skins/skin-contract.css';
+import {useElementDimensions} from './hooks';
 
 import type {DataAttributes} from './utils/types';
+
+type VideoState = 'loading' | 'loaded' | 'playing' | 'paused' | 'error';
+
+type VideoAction = 'play' | 'finishLoad' | 'pause' | 'fail' | 'reset';
+
+const transitions: Record<VideoState, Partial<Record<VideoAction, VideoState>>> = {
+    loading: {
+        play: 'playing',
+        pause: 'paused',
+        fail: 'error',
+        finishLoad: 'loaded',
+    },
+
+    loaded: {
+        play: 'playing',
+        pause: 'paused',
+        reset: 'loading',
+    },
+
+    playing: {
+        pause: 'paused',
+        reset: 'loading',
+    },
+
+    paused: {
+        play: 'playing',
+        reset: 'loading',
+    },
+
+    error: {
+        reset: 'loading',
+    },
+};
+
+const videoReducer = (state: VideoState, action: VideoAction): VideoState =>
+    transitions[state][action] || state;
 
 export type AspectRatio = '1:1' | '16:9' | '4:3';
 
@@ -17,10 +54,16 @@ export const RATIO = {
     '4:3': 4 / 3,
 };
 
-type VideoSource = {
+type VideoSourceWithType = {
     src: string;
     type?: string; // video/webm, video/mp4...
 };
+
+export type VideoSource =
+    | string
+    | ReadonlyArray<string>
+    | VideoSourceWithType
+    | ReadonlyArray<VideoSourceWithType>;
 
 /** Transparent 1x1px PNG  */
 const TRANSPARENT_PIXEL =
@@ -33,13 +76,19 @@ export type VideoProps = {
     /** defaults to 1:1, if both width and height are given, aspectRatio is ignored. To use original video proportions, set aspectRatio to 0 */
     aspectRatio?: AspectRatio | number;
     /** accepts multiple sources */
-    src: string | ReadonlyArray<string> | VideoSource | ReadonlyArray<VideoSource>;
+    src: VideoSource;
     /** defaults to true */
     loop?: boolean;
     /** defaults to true */
     muted?: boolean;
-    /** defaults to true */
-    autoPlay?: boolean;
+    /** defaults to when-loaded. If set to true, behaviour is the same as when the value is equal to when-loaded */
+    autoPlay?: boolean | 'streaming' | 'when-loaded';
+    /** defaults to 10s */
+    loadingTimeout?: number;
+    onError?: () => void;
+    onPlay?: () => void;
+    onPause?: () => void;
+    onLoad?: () => void;
     poster?: string;
     children?: void;
     /** defaults to none */
@@ -52,31 +101,60 @@ const Video = React.forwardRef<HTMLVideoElement, VideoProps>(
         {
             src,
             poster,
-            autoPlay = !isRunningAcceptanceTest(), // default true, but disable autoPlay in screenshot tests
+            autoPlay = 'when-loaded',
             muted = true,
             loop = true,
             preload = 'none',
+            loadingTimeout = 10000,
+            onLoad,
+            onError,
+            onPause,
+            onPlay,
             aspectRatio = '1:1',
             dataAttributes,
             ...props
         },
         ref
     ) => {
-        const borderRadiusContext = useMediaBorderRadius();
+        const [videoStatus, dispatch] = React.useReducer(videoReducer, 'loading');
+        const videoRef = React.useRef<HTMLVideoElement | null>(null);
+        const loadedSource = React.useRef<VideoSource>();
 
+        const borderRadiusContext = useMediaBorderRadius();
         const ratio = typeof aspectRatio === 'number' ? aspectRatio : RATIO[aspectRatio];
 
-        const videoRef = React.useRef<HTMLVideoElement | null>(null);
+        const handleError = React.useCallback(() => {
+            if (videoStatus === 'loading') {
+                dispatch('fail');
+                onError?.();
+            }
+        }, [onError, videoStatus]);
 
         React.useEffect(() => {
+            if (loadedSource.current !== src) {
+                loadedSource.current = src;
+                const loadingTimeoutId = setTimeout(handleError, loadingTimeout);
+                videoRef.current?.load();
+
+                return () => {
+                    clearTimeout(loadingTimeoutId);
+                };
+            }
+        }, [src, loadingTimeout, handleError]);
+
+        const handleLoadFinish = () => {
+            onLoad?.();
             const video = videoRef.current;
-            if (video && autoPlay && video.paused) {
+            const shouldAutoPlay = autoPlay && !isRunningAcceptanceTest();
+
+            dispatch('finishLoad');
+            if (video && shouldAutoPlay && video.paused) {
                 video.play();
             }
-        }, [autoPlay]);
+        };
 
         // normalize sources
-        const sources: Array<VideoSource> = (Array.isArray(src) ? src : [src]).map((source) => {
+        const sources: Array<VideoSourceWithType> = (Array.isArray(src) ? src : [src]).map((source) => {
             if (typeof source === 'string') {
                 return {src: source};
             } else {
@@ -84,23 +162,47 @@ const Video = React.forwardRef<HTMLVideoElement, VideoProps>(
             }
         });
 
+        const showPoster = videoStatus === 'error' || videoStatus === 'loading' || videoStatus === 'loaded';
+        const {ref: posterRef, width: posterWidth, height: posterHeight} = useElementDimensions();
+
         const video = (
             <video
                 ref={combineRefs(ref, videoRef)}
                 playsInline
                 disablePictureInPicture
                 disableRemotePlayback
-                autoPlay={autoPlay}
                 muted={muted}
                 loop={loop}
                 className={styles.video}
                 preload={preload}
+                onLoadStart={() => {
+                    dispatch('reset');
+                }}
+                onError={handleError}
+                onPause={() => {
+                    onPause?.();
+                    dispatch('pause');
+                }}
+                onPlay={() => {
+                    onPlay?.();
+                    dispatch('play');
+                }}
+                onCanPlay={() => {
+                    if (autoPlay === 'streaming') handleLoadFinish();
+                }}
+                onCanPlayThrough={() => {
+                    if (autoPlay !== 'streaming') handleLoadFinish();
+                }}
                 // This transparent pixel fallback avoids showing the ugly "play" image in android webviews
-                poster={poster || TRANSPARENT_PIXEL}
+                poster={TRANSPARENT_PIXEL}
                 {...getPrefixedDataAttributes(dataAttributes)}
                 style={{
                     // For some reason adding this style with classnames doesn't add the border radius in safari
                     borderRadius: !borderRadiusContext ? 0 : vars.borderRadii.container,
+                    visibility: showPoster ? 'hidden' : 'visible',
+                    position: showPoster || ratio !== 0 ? 'absolute' : 'static',
+                    width: showPoster ? posterWidth : '100%',
+                    height: showPoster ? posterHeight : '100%',
                 }}
             >
                 {sources.map(({src, type}, index) => (
@@ -109,28 +211,35 @@ const Video = React.forwardRef<HTMLVideoElement, VideoProps>(
             </video>
         );
 
+        const withErrorFallback = !!(ratio !== 0 || (props.width && props.height));
+
+        const posterImage = poster ? (
+            <ImageContent
+                ref={posterRef}
+                aspectRatio={aspectRatio}
+                width={props.width}
+                height={props.height}
+                src={poster}
+            />
+        ) : withErrorFallback ? (
+            <div style={{position: 'absolute', width: '100%', height: '100%'}}>
+                <ImageError
+                    ref={posterRef}
+                    noBorderRadius={!borderRadiusContext}
+                    withIcon={videoStatus === 'error'}
+                />
+            </div>
+        ) : undefined;
+
         return (
-            <AspectRatioElement aspectRatio={ratio} width={props.width} height={props.height}>
-                {isSafari() ? (
-                    <div
-                        style={{
-                            width: '100%',
-                            height: '100%',
-                            /**
-                             * In safari, when using a video with poster, the transition from pause to play does a flicker,
-                             * To avoid this, in Safari browsers, instead of using the poster attribute, we use a
-                             * wrapper with the poster as background image
-                             */
-                            backgroundImage: poster ? `url("${poster}")` : undefined,
-                            backgroundSize: 'cover',
-                            backgroundPosition: '50% 50%',
-                        }}
-                    >
-                        {video}
-                    </div>
-                ) : (
-                    video
-                )}
+            <AspectRatioElement
+                style={{position: 'relative'}}
+                aspectRatio={ratio}
+                width={props.width}
+                height={props.height}
+            >
+                {video}
+                {showPoster && posterImage}
             </AspectRatioElement>
         );
     }